+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 js-cropbox
 ==========
 
+0.9.2 [2017-05-05]
+------------------
+
+- Added method `getVariants()`.
+- Fixed first argument of `constructor()`.
+- Added tests for `constructor()` and `getVariants()` methods.
+- Modified `README.md`.
+
 0.9.1 [2017-05-02]
 ------------------